requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Flask==2.0.3
Flask-RESTful==0.3.9
fonttools==4.29.1
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
pycatch22==0.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==3.1.0
urllib3==1.26.8
Werkzeug==2.0.3
